requirements.txt
# Existing
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
openai
python-dotenv
pydantic
pydantic-settings
requests

# New for MVP Phase 1
celery[redis]
redis
cryptography
python-jose[cryptography]
passlib[bcrypt]
httpx
alembic
